GUARD-1723: SystemStatusEnvironment.wp_memory_limit field type modified from int to long to support of values more than 2GB (2147483648)

+++ WooCommerce/v2/SystemStatus.cs
@@ -125,7 +125,7 @@ namespace WooCommerceNET.WooCommerce.v2
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? wp_memory_limit { get; set; }
+        public long? wp_memory_limit { get; set; }
 
         /// <summary>
         /// Is WordPress debug mode active?
